Extract _kometa_paths.py from _legacy.py
Moves the three install-mode-aware Kometa path resolvers into their own
submodule:

- get_kometa_root_path (~50 lines): resolves the Kometa root folder,
  honoring app.config, session, persisted 900-kometa runtime section,
  and the managed default under CONFIG_DIR/kometa
- get_kometa_config_dir (~30 lines): resolves the config directory,
  handling external-install overrides
- get_kometa_log_dir (~35 lines): resolves the log directory, deriving
  from config_dir when no explicit override is set

Verbatim moves. The three functions are chained (log -> config -> root)
and now share a private module namespace so the chained calls stay
unqualified.

_legacy.py's save_to_named_config still calls get_kometa_root_path and
get_kometa_config_dir; wired via a top-level import from _kometa_paths.
No import cycle: _kometa_paths only imports from _install_mode inside
functions (lazy), and _install_mode's top-level import of CONFIG_DIR
from _legacy resolves fine.

External call sites unchanged (40+ across quickstart.py, modules/*, and
tests) — all reach these via helpers.foo() and the star re-export in
helpers/__init__.py.

Reduces _legacy.py from 277 to 156 lines (-121). All 1077 tests pass.
